requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
spacy==3.7.2
PyPDF2==3.0.1
python-docx==1.1.0
pymongo[srv]
dnspython
python-dotenv==1.0.0
google-genai
